requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
PyYAML>=6.0
tqdm>=4.65.0

# ── Finance & Data ────────────────────────────────────────────────────────────
yfinance>=0.2.28
ta>=0.10.2               # Technical-analysis indicators

# ── ML / Deep Learning ────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# ── Image ─────────────────────────────────────────────────────────────────────
Pillow>=10.0.0
scikit-image>=0.21.0
mplfinance>=0.12.9b7     # Candlestick chart generation

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Notebook ──────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
